пытаюсь реализовать передачу данных из сервиса

+++ app/src/main/java/com/orishkevich/accelerometerapp/model/AccelModel.java
@@ -40,7 +40,7 @@ public class AccelModel {
         this.z = z;
     }
 
-    public float getMil() {
+    public long getMil() {
         return mil;
     }
